normalize heuristic

+++ src/main/java/agents/mcts/Stats.java
@@ -5,13 +5,13 @@ package main.java.agents.mcts;
  */
 public class Stats {
     protected int numPlays; // number of times this action/edge has been played
-    protected int totalReward; // aggregate reward seen along edge
+    protected double totalReward; // aggregate reward seen along edge
 
     /**
      * Initialize Stats object with zero values
      */
     public Stats() {
         this.numPlays = 0;
-        this.totalReward = 0;
+        this.totalReward = 0.0;
     }
 }
